Some script cleanup and reorganizing
