+++ gui/qt/deploy-scripts/BuildQt56StaticDevArchive.py
@@ -1,11 +1,12 @@
 import os
 import sys
 import fnmatch
-import errno
 import subprocess
 import shutil
 import re
 import glob
+
+from util import *
 
 SEVENZIP="C:\\Program Files\\7-Zip\\7z.exe"
 QTBASEDIR="C:\\"
@@ -26,49 +27,6 @@ QT_LIB_INCLUDE = "core gui quick widgets quickwidgets network qml"
 # Binary excludes
 # (Don't change unless you know what you're doing!)
 BIN_EXCLUDES = [ "assistant.exe", "designer.exe", "idc.exe", "idc.pdb", "linguist.exe", "moc.pdb", "pixeltool.exe", "qdbus*", "rcc.pdb", "testcon*", "windeployqt*"]
-
-def silentremove(filename):
-    try:
-        os.remove(filename)
-    except OSError as e: # this would be "except OSError, e:" before Python 2.6
-        if e.errno != errno.ENOENT: # errno.ENOENT = no such file or directory
-            raise # re-raise exception if a different error occured
-
-def simple_exec(cmd):
-    print("   -> Executing command: %s" % " ".join(cmd))
-    
-    retcode = subprocess.call(cmd)
-    
-    if retcode != 0:
-        print("   !! ERROR: Command returned exit code %i!" % retcode)
-        print("   !!        Command: %s" % " ".join(cmd))
-        return False
-    
-    return True
-
-def silent_exec(cmd):
-    print("   -> Executing command: %s" % " ".join(cmd))
-    
-    FNULL = open(os.devnull, 'w')
-    retcode = subprocess.call(cmd, stdout=FNULL, stderr=subprocess.STDOUT)
-    
-    if retcode != 0:
-        print("   !! ERROR: Command returned exit code %i!" % retcode)
-        print("   !!        Command: %s" % " ".join(cmd))
-        return False
-    
-    return True
-
-# tzot @ StackOverflow:
-# http://stackoverflow.com/a/600612
-def mkdir_p(path):
-    try:
-        os.makedirs(path)
-    except OSError as exc:  # Python >2.5
-        if exc.errno == errno.EEXIST and os.path.isdir(path):
-            pass
-        else:
-            raise
 
 def collect_qt_static_files(proj_file):
     proj_fh = open(proj_file, "r")
@@ -160,7 +118,8 @@ print(" * Stage 1: Analyzing CEmu")
 mkdir_p("build_32")
 
 os.chdir("build_32")
-if not simple_exec([r'C:\Qt\Qt5.6.0-static\bin\qmake', '-spec', 'win32-msvc2015', '-tp', 'vc', r'"..\..\CEmu.pro"']):
+print(os.getcwd())
+if not simple_exec([r'C:\Qt\Qt5.6.0-static\bin\qmake', '-spec', 'win32-msvc2015', '-tp', 'vc', r'..\..\CEmu.pro']):
     print(" ! ERROR: Creating project files for x86 failed!")
     sys.exit(1)
 os.chdir("..")
